+++ frontend/src/components/Mesh/TestModal.tsx
@@ -1,11 +1,9 @@
 import * as React from 'react';
 import { TracingCheck, TracingInfo } from '../../types/TracingInfo';
-import { Button, ButtonVariant, Modal, ModalVariant, Popover, PopoverPosition, Tab } from '@patternfly/react-core';
+import { Button, Modal, ModalVariant, Tab } from '@patternfly/react-core';
 import { kialiStyle } from '../../styles/StyleUtils';
 import { useKialiTranslation } from '../../utils/I18nUtils';
 import { PFColors } from '../Pf/PfColors';
-import { KialiIcon } from '../../config/KialiIcon';
-import * as API from '../../services/Api';
 import { ExternalServiceInfo, TempoUrlFormat } from '../../types/StatusState';
 import { isParentKiosk } from '../Kiosk/KioskActions';
 import { isTempoService, TempoUrlProvider } from '../../utils/tracing/UrlProviders/Tempo';
@@ -15,16 +13,11 @@ import { KialiDispatch } from '../../types/Redux';
 import { bindActionCreators } from 'redux';
 import { TracingActions } from '../../actions/TracingActions';
 import { connect } from 'react-redux';
-import AceEditor from 'react-ace';
-import { getKialiTheme } from 'utils/ThemeUtils';
-import { Theme } from '../../types/Common';
-import { istioAceEditorStyle } from '../../styles/AceEditorStyle';
-import { aceOptions, yamlDumpOptions } from '../../types/IstioConfigDetails';
-import ReactAce from 'react-ace/lib/ace';
 import { classes } from 'typestyle';
 import { basicTabStyle } from '../../styles/TabStyles';
 import { ParameterizedTabs } from '../Tab/Tabs';
-import { dump } from 'js-yaml';
+import { CheckConfig } from './CheckConfig';
+import { TestConfig } from './TestConfig';
 
 type ReduxProps = {
   externalServices: ExternalServiceInfo[];
@@ -48,27 +41,6 @@ const modalStyle = kialiStyle({
   overflowY: 'hidden'
 });
 
-const containerStyle = kialiStyle({
-  backgroundColor: PFColors.Black1000,
-  color: PFColors.Blue100,
-  fontFamily: 'Courier New, Courier, monospace',
-  margin: 0,
-  padding: '0.5em',
-  resize: 'none',
-  whiteSpace: 'pre',
-  width: '100%',
-  overflowX: 'scroll'
-});
-
-const blueDisplay = kialiStyle({
-  color: PFColors.Blue400
-});
-
-const blueDarkDisplay = kialiStyle({
-  color: PFColors.Blue200,
-  padding: '0 0.5em'
-});
-
 const tabStyle = kialiStyle({
   $nest: {
     '&& .pf-v5-c-tabs__list': {
@@ -77,7 +49,7 @@ const tabStyle = kialiStyle({
   }
 });
 
-const helpStyle = kialiStyle({
+export const helpStyle = kialiStyle({
   marginBottom: '0.6em',
   marginLeft: '0.375rem',
   $nest: {
@@ -88,16 +60,6 @@ const helpStyle = kialiStyle({
   }
 });
 
-const configStyle = kialiStyle({
-  fontFamily: 'Courier New, Courier, monospace',
-  fontSize: 'small',
-  margin: '1.25em'
-});
-
-const blockDisplay = kialiStyle({
-  display: 'block'
-});
-
 const defaultTab = 'checkConfig';
 
 const tabIndex: { [tab: string]: number } = {
@@ -105,7 +67,7 @@ const tabIndex: { [tab: string]: number } = {
   testConfig: 1
 };
 
-const validateExternalUrl = (
+export const validateExternalUrl = (
   externalServices: ExternalServiceInfo[],
   kiosk: string,
   info?: TracingInfo
@@ -142,202 +104,23 @@ const validateExternalUrl = (
 
 export const TestModalComp: React.FC<TestModalProps> = (props: TestModalProps) => {
   const { t } = useKialiTranslation();
-  const [loading, setLoading] = React.useState(false);
-  const [error, setError] = React.useState<string | null>(null);
-  const labels = {
-    namespaceSelector: 'namespace_selector',
-    provider: 'provider',
-    url: 'internal_url',
-    useGRPC: 'use_grpc'
-  };
-  const externalUrl = validateExternalUrl(props.externalServices, props.kiosk, props.tracingInfo);
-  const aceEditorRef = React.useRef<ReactAce | null>(null);
+
   const [currentTab, setCurrentTab] = React.useState(defaultTab);
 
   const renderTabs = (): React.ReactNode[] => {
     const checkConfig = (
       <Tab eventKey={0} title={t('Check Config')} key="checkConfig">
-        <div style={{ paddingTop: '0.5em' }}>
-          <div>
-            <span style={{ paddingTop: '0.25em' }}>
-              {t('Check for possible configuration(s) of external_services.tracing')}
-            </span>
-            <Popover
-              data-test="check-status-help"
-              position={PopoverPosition.auto}
-              headerContent={
-                <div>
-                  <span>{t('Check Status')}</span>
-                </div>
-              }
-              bodyContent={
-                <>
-                  {t(
-                    'Check the usual ports for the tracing service and provide a subset of the tracing configuration based on the tracing services found for external_services.tracing.'
-                  )}
-                  <br />
-                  {t(
-                    'While the health check is based on whether the URL response returns an HTTP 200, the services check performs a more exhaustive verification by attempting to analyze if the traces response is correct. It is important that internal_url is defined, as it relies on this host to perform the checks. When in_cluster config is false, it will use the external_url'
-                  )}
-                </>
-              }
-            >
-              <KialiIcon.Help className={helpStyle} />
-            </Popover>
-          </div>
-          <div style={{ display: 'flex' }}>
-            <Button onClick={handleCheckService} disabled={loading} variant={ButtonVariant.secondary}>
-              {loading ? t('Verifying...') : t('Check Status')}
-            </Button>
-            {props.tracingDiagnose && (
-              <span style={{ position: 'absolute', right: '2.4em' }}>
-                <Button
-                  variant={ButtonVariant.link}
-                  aria-label={t('Close')}
-                  isInline
-                  onClick={() => props.setTracingDiagnose()}
-                >
-                  <KialiIcon.Close />
-                </Button>
-              </span>
-            )}
-          </div>
-          {props.tracingDiagnose && <span style={{ color: 'green' }}>{props.tracingDiagnose.message}</span>}
-          {error && (
-            <div>
-              <span style={{ color: 'red' }}>{error}</span>
-            </div>
-          )}
-          {props.tracingDiagnose?.validConfig && (
-            <>
-              <div style={{ margin: '0.5em 0', display: 'flex' }}>
-                <span>
-                  {props.tracingDiagnose?.validConfig.length === 0 && (
-                    <>No configurations found. See logs for details</>
-                  )}
-                </span>
-              </div>
-              <div>
-                {props.tracingDiagnose?.validConfig?.map((item, i) => (
-                  <>
-                    <div className={configStyle}>
-                      {Object.keys(item).map(key => {
-                        if (labels[key] !== undefined && item[key] != null) {
-                          return (
-                            <span className={blockDisplay}>
-                              <span className={blueDisplay}>{labels[key]}:</span> {item[key].toString()}
-                            </span>
-                          );
-                        }
-                        return null;
-                      })}
-                      {item.warning && <span style={{ color: 'red' }}>{item.warning}</span>}
-                    </div>
-                    {props.tracingDiagnose?.validConfig && i < props.tracingDiagnose?.validConfig?.length - 1 && <hr />}
-                  </>
-                ))}
-              </div>
-            </>
-          )}
-          {props.tracingDiagnose?.logLine && externalUrl && (
-            <div>
-              <span className={configStyle}>{externalUrl}</span>
-            </div>
-          )}
-          {props.tracingDiagnose?.logLine && (
-            <div className={containerStyle}>
-              {props.tracingDiagnose.logLine.map(log => (
-                <>
-                  <div>
-                    <span>
-                      <span className={blueDisplay}>{log.time.substring(0, 19)}</span>
-                      <span className={blueDarkDisplay}>[{log.test}]</span>
-                      {log.result}
-                    </span>
-                  </div>
-                </>
-              ))}
-            </div>
-          )}
-        </div>
+        <CheckConfig cluster={props.cluster} />
       </Tab>
     );
 
-    const theme = getKialiTheme();
-
-    const getConfigToString = dump(props.tracingInfo, {
-      noRefs: true,
-      skipInvalid: true,
-      ...yamlDumpOptions
-    });
-
     const testConfig = (
       <Tab eventKey={1} title={t('Test Configuration')} key="testConfig">
-        <div style={{ paddingTop: '10px' }}>
-          <span>
-            {t('external_services.tracing configuration:')}{' '}
-            <Popover
-              data-test="check-status-help"
-              position={PopoverPosition.auto}
-              headerContent={
-                <div>
-                  <span>Check Status</span>
-                </div>
-              }
-              bodyContent={
-                <>
-                  {t(
-                    'Check the usual ports for the tracing service and provide a subset of the tracing configuration based on the tracing services found for external_services.tracing.'
-                  )}
-                  <br />
-                  {t(
-                    'While the health check is based on whether the URL response returns an HTTP 200, the services check performs a more exhaustive verification by attempting to analyze if the traces response is correct. It is important that internal_url is defined, as it relies on this host to perform the checks. When in_cluster config is false, it will use the external_url'
-                  )}
-                </>
-              }
-            >
-              <KialiIcon.Help className={helpStyle} />
-            </Popover>
-          </span>
-          <AceEditor
-            ref={aceEditorRef}
-            mode="yaml"
-            theme={theme === Theme.DARK ? 'twilight' : 'eclipse'}
-            width="100%"
-            className={istioAceEditorStyle}
-            wrapEnabled={true}
-            readOnly={false}
-            setOptions={aceOptions ?? { foldStyle: 'markbegin' }}
-            value={getConfigToString}
-          />
-          <Button style={{ marginTop: '10px' }} variant={ButtonVariant.secondary}>
-            Test Config
-          </Button>
-        </div>
+        <TestConfig tracingInfo={props.tracingInfo} />
       </Tab>
     );
 
     return [checkConfig, testConfig];
-  };
-
-  const fetchCheckService = async (): Promise<void> => {
-    setLoading(true);
-    props.setTracingDiagnose();
-    setError(null);
-
-    return API.getDiagnoseStatus(props.cluster)
-      .then(response => {
-        props.setTracingDiagnose(response.data);
-        setLoading(false);
-      })
-      .catch(err => {
-        setLoading(false);
-        setError(`Could not fetch diagnose info ${err}`);
-      });
-  };
-
-  const handleCheckService = async (): Promise<void> => {
-    fetchCheckService();
   };
 
   if (!props.isOpen) {
